Add fabs to math.h

+++ src/libk/math.c
@@ -1,6 +1,6 @@
 #include <math.h>
 
-double abs(double x) {
+int abs(int x) {
     return x < 0 ? -x : x;
 }
